ai-engine-python/requirements.txt
fastapi
uvicorn
openai
python-dotenv
huggingface-hub